+++ packages/angular-workspace/spright-angular/CHANGELOG.md
@@ -1,8 +1,40 @@
 # Change Log - @ni/spright-angular
 
-This log was last generated on Wed, 24 Jul 2024 16:11:28 GMT and should not be manually modified.
+This log was last generated on Tue, 30 Jul 2024 20:21:00 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 3.0.7
+
+Tue, 30 Jul 2024 20:21:00 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v2.0.15
+
+## 3.0.6
+
+Tue, 30 Jul 2024 18:17:01 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v2.0.14
+
+## 3.0.5
+
+Tue, 30 Jul 2024 16:53:07 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v2.0.13
+
+## 3.0.4
+
+Tue, 30 Jul 2024 14:34:23 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v2.0.12
 
 ## 3.0.3